Group tests of invalid number of octets
After think about it I think that we have 2 options:
1. Split the tests between less than and more than cases.
2. Group all tests of invalid number of octets

In this problem, in my opinion, it's clear. As IP v4 are not going to change its format, group all this cases in one test it's enough.
But in other cases where the domain can change I think will be better to split them in 2 different tests in order to:
* Had to change only one of them if the limit changes (4 octets to 4-6 octets)
* Specify the limits clearly