Add case of invalid character in third octet. Also added a case of the plus character
We detected that the third character force us to change our current implementation and our focus. Currently we are trying to achieve a validation on the four octets, maybe not a perfect validation but we want to achieve the array step. So, we disable the plus character test case we pick an easiest case to keep moving and "document" our problematic test case to work on it after